Release: Update NGC versions post-24.08 release (#7565)

+++ qa/common/util.sh
@@ -257,7 +257,7 @@ function run_server_nowait () {
         return
     fi
 
-    if [[ "$(< /proc/sys/kernel/osrelease)" == *microsoft* ]]; then
+    if [[ -v WSL_DISTRO_NAME ]] || [[ -v MSYSTEM ]]; then
         # LD_PRELOAD not yet supported on windows
         if [ -z "$SERVER_LD_PRELOAD" ]; then
             echo "=== Running $SERVER $SERVER_ARGS"
@@ -329,7 +329,7 @@ function kill_server () {
     # causes the entire WSL shell to just exit. So instead we must use
     # taskkill.exe which can only forcefully kill tritonserver which
     # means that it does not gracefully exit.
-    if [[ "$(< /proc/sys/kernel/osrelease)" == *microsoft* ]]; then
+    if [[ -v WSL_DISTRO_NAME ]]; then
         # Disable -x as it makes output below hard to read
         oldstate="$(set +o)"; [[ -o errexit ]] && oldstate="$oldstate; set -e"
         set +x
@@ -353,6 +353,8 @@ function kill_server () {
         fi
 
         set +vx; eval "$oldstate"
+    elif [[ -v MSYSTEM ]] ; then
+        taskkill //F //IM tritonserver.exe
     else
         # Non-windows...
         kill $SERVER_PID
@@ -512,17 +514,23 @@ remove_array_outliers() {
 
 function setup_virtualenv() {
     # Create and activate virtual environment
-    virtualenv --system-site-packages venv
-    source venv/bin/activate
-    pip install pytest
+    if [[ -v MSYSTEM ]]; then
+      pip3 install pytest
+    else
+      virtualenv --system-site-packages venv
+      source venv/bin/activate
+      pip install pytest
+    fi
 
     if [[ ${TEST_WINDOWS} == 1 ]]; then
-        pip3 install "numpy<2" tritonclient[all]
+      pip3 install "numpy<2" tritonclient[all]
     fi
 }
 
 function deactivate_virtualenv() {
     # Deactivate virtual environment and clean up
+  if [[ ! -v MSYSTEM ]]; then
     deactivate
     rm -fr venv
+  fi
 }
